Sessions: fix for a memory corruption issue

+++ plugins/Sessions/Src/stdafx.h
@@ -70,7 +70,7 @@ int SaveSessionDate();
 extern MCONTACT session_list_recovered[255];
 extern MCONTACT session_list[255];
 extern int g_ses_limit;
-extern size_t g_ses_count;
+extern int g_ses_count;
 extern bool g_bExclHidden;	
 extern bool g_bWarnOnHidden;
 extern bool g_bOtherWarnings;
